:pencil2: Type 지정 위치 변경
